fix(tmux): fix flicker in status bar
A bit of an interesting bug. Until a command finishes, tmux will display
a placeholder value in the status bar. Since the `date` command is
running every second, the placeholder message causes a flicker. To
mitigate this issue, tmux caches the output from the last invocation,
but only if the command *itself* hasn't changed.

It turns out that tmux runs the status line through `strftime` before
executing any commands. So, in addition to the `date` invocation being
unnecessary, the actual command being invoked was changing every time,
defeating the caching!

So, the fix is to just use the `strftime` formatting.